server/requirements.txt
selenium
undetected_chromedriver
langchain-text-splitters
langchain-google-genai
google-generativeai
python-dotenv